add list test, fix list construction
There are two issues fixed this commit: the first one is the
kv_list_delete function, that did not actually delete correctly when
the node was at the end of the list; the second is a much needed sanity
check in the kv_list_add function, that now sets to NULL the next
pointer of the newly inserted node.

The list test is simple: it builds a lists, then removes certain
elements, and checks if the resulting list is correct.